+++ src/application.js
@@ -1,41 +1,73 @@
 "use strict";
 
-import restify from "restify";
+import path from "path";
 
 import _ from "lodash";
+import callsite from "callsite";
+import restify from "restify";
+
 import Loader from "./loader";
+import ModelManager from "./model_manager";
 import Router from "./router";
 
+const DEFAULT_CONNECTION_SETTINGS = {
+  dialect: "sqlite",
+  database: "test",
+  username: "test",
+  password: "test"
+};
+const DEFAULT_LISTEN_PORT = 3000;
 const DEFAULT_MIDDLEWARES = [
   restify.gzipResponse(),
+
+  /* eslint-disable new-cap */
   restify.CORS(),
+
+  /* eslint-enable new-cap */
   restify.authorizationParser(),
   restify.bodyParser(),
   restify.fullResponse(),
   restify.queryParser()
 ];
 
-/**
- * TODO:
- *   - add DEFAULT_CONTROLLER_LOOKUP_PATH
- *     - should be <caller location>/contrllers
- */
-
 class Application {
   constructor(options = {}) {
-    const serverOptions = options.server || {};
+    // who are you
+    const caller = callsite()[1].getFileName();
+    const callerDirectory = path.dirname(caller);
+    const DEFAULT_CONTROLLER_LOOKUP_PATH = path.resolve(callerDirectory, "controllers");
+    const DEFAULT_MODEL_LOOKUP_PATH = path.resolve(callerDirectory, "models");
 
-    // TODO: merge with defaults
-    const middlewares = _.union(DEFAULT_MIDDLEWARES, serverOptions.middlewares || []);
-    const app = options.app || restify.createServer(serverOptions);
+    options.controllers = options.controllers || {};
+    options.controllers.dir = options.controllers.dir || DEFAULT_CONTROLLER_LOOKUP_PATH;
+    options.database = options.database || {};
+    options.database.connection = options.database.connection || DEFAULT_CONNECTION_SETTINGS;
+    options.database.models = options.database.models || {};
+    options.database.models.dir = options.database.models.dir || DEFAULT_MODEL_LOOKUP_PATH;
+    options.server = options.server || {};
+    options.server.middlewares = options.server.middlewares || [];
+
+    const app = options.app || restify.createServer(options.server);
+    const connection = options.database.connection;
     const controllerLoader = new Loader({
       type: "controller",
-      path: options.controllers.dir // || DEFAULT_CONTROLLER_LOOKUP_PATH
+      path: options.controllers.dir
     });
+    const middlewares = _.union(DEFAULT_MIDDLEWARES, options.server.middlewares);
+
+    this._internalModels = new Loader({
+      type: "model",
+      path: options.database.models.dir
+    });
+    this.modelManager = new ModelManager({ connection });
+    this._addModels();
+    this._associateModels();
+
     const routerSettings = {
       app,
       loader: {
-        controllers: controllerLoader
+        controllers: controllerLoader,
+        models: this.modelManager.models
       }
     };
 
@@ -49,9 +81,27 @@ class Application {
     return this.app;
   }
 
-  start(port = 3000) {
+  start(port = DEFAULT_LISTEN_PORT) {
     return new Promise((resolve, reject) => {
-      this.app.listen(port, function () { resolve(); })
+      this.app.listen(port, () => { resolve(); });
+    });
+  }
+
+  _addModels() {
+    const _internalModels = this._internalModels;
+
+    Object.keys(_internalModels.modules).forEach(model => {
+      this.modelManager.addModel(_internalModels.modules[model]);
+    });
+  }
+
+  _associateModels() {
+    const modelManager = this.modelManager;
+
+    Object.keys(modelManager.models).forEach(model => {
+      if (modelManager.models[model].associate) {
+        modelManager.models[model].associate(modelManager.models[model], modelManager.models);
+      }
     });
   }
 }