+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+<a name="0.0.5"></a>
+## [0.0.5](https://github.com/dylanfoster/parch/compare/0.0.4...v0.0.5) (2016-06-22)
+
+
+### Features
+
+* Add Sequelize support
+* **application:** add default lookup paths for controllers and models and clean up constructor ([cbb5985](https://github.com/dylanfoster/parch/commit/cbb5985))
+* **application:** clean up constructor some more ([3432f07](https://github.com/dylanfoster/parch/commit/3432f07))
+* **models:** add base model class and pending tests ([2957baf](https://github.com/dylanfoster/parch/commit/2957baf))
+* **models:** add model managaer ([7da0e24](https://github.com/dylanfoster/parch/commit/7da0e24))
+
+
+
 <a name="0.0.4"></a>
 ## [0.0.4](https://github.com/dylanfoster/parch/compare/0.0.3...v0.0.4) (2016-06-16)